[Http] Move routing middleware from Middleware to Routing.
