+++ src/Valkyrja/Http/Routing/Middleware/RouteMatched/RequestStructMiddleware.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Valkyrja\Http\Routing\Middleware;
+namespace Valkyrja\Http\Routing\Middleware\RouteMatched;
 
 use Override;
 use Valkyrja\Http\Message\Enum\StatusCode;
