+++ tests/Unit/Application/Kernel/ApplicationTest.php
@@ -51,7 +51,6 @@ use Valkyrja\Event\Collector\Contract\CollectorContract as EventCollector;
 use Valkyrja\Event\Data\Data as EventData;
 use Valkyrja\Event\Dispatcher\Contract\DispatcherContract as EventDispatcher;
 use Valkyrja\Http\Message\Factory\Contract\ResponseFactoryContract;
-use Valkyrja\Http\Middleware\Cache\CacheResponseMiddleware;
 use Valkyrja\Http\Middleware\Handler\Contract\RequestReceivedHandlerContract;
 use Valkyrja\Http\Middleware\Handler\Contract\RouteDispatchedHandlerContract as HttpRouteDispatchedHandlerContract;
 use Valkyrja\Http\Middleware\Handler\Contract\RouteMatchedHandlerContract as HttpRouteMatchedHandlerContract;
@@ -66,9 +65,10 @@ use Valkyrja\Http\Routing\Data\Data as HttpData;
 use Valkyrja\Http\Routing\Dispatcher\Contract\RouterContract as HttpRoutingRouter;
 use Valkyrja\Http\Routing\Factory\Contract\ResponseFactoryContract as HttpRoutingResponseFactory;
 use Valkyrja\Http\Routing\Matcher\Contract\MatcherContract;
-use Valkyrja\Http\Routing\Middleware\RequestStructMiddleware;
-use Valkyrja\Http\Routing\Middleware\ResponseStructMiddleware;
-use Valkyrja\Http\Routing\Middleware\ViewRouteNotMatchedMiddleware;
+use Valkyrja\Http\Routing\Middleware\CacheResponseMiddleware;
+use Valkyrja\Http\Routing\Middleware\RouteMatched\RequestStructMiddleware;
+use Valkyrja\Http\Routing\Middleware\RouteMatched\ResponseStructMiddleware;
+use Valkyrja\Http\Routing\Middleware\RouteNotMatched\ViewRouteNotMatchedMiddleware;
 use Valkyrja\Http\Routing\Processor\Contract\ProcessorContract;
 use Valkyrja\Http\Routing\Url\Contract\UrlContract;
 use Valkyrja\Http\Server\Handler\Contract\RequestHandlerContract;
